refactor(client-core): extract outbound SASL handshake from broker
Move the transport-agnostic outbound SASL state machine
(outbound_sasl + PLAIN/SCRAM/GSSAPI client helpers + framing) out of
crates/broker/src/network/client.rs into crates/client-core/src/sasl.rs
as a reusable handshake over any AsyncRead + AsyncWrite stream. The
broker's InterBrokerClient now maps its InterBrokerCredentials onto the
client-core SaslCredentials and calls the shared fn; inter-broker auth
semantics are unchanged (raft_sasl regression passes).
