+++ crates/broker/src/network/client.rs
@@ -5,36 +5,52 @@
 //! Used by the replicator's Fetch path, the raft transport's
 //! outbound dial, and the controller-heartbeat loop.
 
-use bytes::{Buf, BufMut, BytesMut};
-use crabka_protocol::owned::sasl_authenticate_request::SaslAuthenticateRequest;
-use crabka_protocol::owned::sasl_authenticate_response::SaslAuthenticateResponse;
-use crabka_protocol::owned::sasl_handshake_request::SaslHandshakeRequest;
-use crabka_protocol::owned::sasl_handshake_response::SaslHandshakeResponse;
-use crabka_protocol::{Decode, Encode};
-use crabka_security::{ListenerProtocol, SaslMechanism, ScramClientExchange};
+use crabka_security::ListenerProtocol;
 use thiserror::Error;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
+use tokio::io::{AsyncRead, AsyncWrite};
 use tokio::net::TcpStream;
 use tokio_rustls::TlsConnector;
 
-use std::path::Path;
 use std::sync::Arc;
 
 use crate::config::InterBrokerCredentials;
 
-/// API keys this client speaks directly (everything else is the caller's
-/// problem — we return a raw stream once auth is done).
-const API_KEY_SASL_HANDSHAKE: i16 = 17;
-const API_KEY_SASL_AUTHENTICATE: i16 = 36;
-
-/// `client_id` advertised in outbound Kafka request headers. Visible in
-/// broker logs as the connection's reporter id.
-const OUTBOUND_CLIENT_ID: &str = "crabka-inter-broker";
-
-/// `max_recv_size` advertised in the client's RFC 4752 security-layer choice.
-/// Auth-only QOP means no data is wrapped post-handshake, so the value only
-/// needs to be a sane non-zero buffer; mirror the server's offer size.
-const GSSAPI_MAX_RECV_SIZE: u32 = 0x1_0000;
+/// Map the broker's [`InterBrokerCredentials`] onto the client-core
+/// [`crabka_client_core::SaslCredentials`] understood by the shared
+/// [`crabka_client_core::outbound_sasl`] handshake. The two enums carry
+/// the same variants; this is a field-for-field copy. Shared with the
+/// RLMM bootstrap so the dialer and the metadata client agree on the
+/// mapping.
+pub(crate) fn to_client_creds(c: &InterBrokerCredentials) -> crabka_client_core::SaslCredentials {
+    match c {
+        InterBrokerCredentials::Plain { username, password } => {
+            crabka_client_core::SaslCredentials::Plain {
+                username: username.clone(),
+                password: password.clone(),
+            }
+        }
+        InterBrokerCredentials::Scram {
+            mechanism,
+            username,
+            password,
+        } => crabka_client_core::SaslCredentials::Scram {
+            mechanism: *mechanism,
+            username: username.clone(),
+            password: password.clone(),
+        },
+        InterBrokerCredentials::Gssapi {
+            keytab_path,
+            client_principal,
+            service_name,
+            kdc_url,
+        } => crabka_client_core::SaslCredentials::Gssapi {
+            keytab_path: keytab_path.clone(),
+            client_principal: client_principal.clone(),
+            service_name: service_name.clone(),
+            kdc_url: kdc_url.clone(),
+        },
+    }
+}
 
 #[derive(Debug, Error)]
 pub enum InterBrokerError {
@@ -107,7 +123,9 @@ impl InterBrokerClient {
             let creds = self.creds.clone().ok_or_else(|| {
                 InterBrokerError::Config("SASL listener without inter_broker_credentials".into())
             })?;
-            run_outbound_sasl(&mut *stream, &creds, server_name).await?;
+            crabka_client_core::outbound_sasl(&mut *stream, &to_client_creds(&creds), server_name)
+                .await
+                .map_err(|e| InterBrokerError::Sasl(e.to_string()))?;
         }
         Ok(stream)
     }
@@ -151,334 +169,14 @@ impl InterBrokerClient {
             let creds = self.creds.clone().ok_or_else(|| {
                 InterBrokerError::Config("SASL listener without inter_broker_credentials".into())
             })?;
-            run_outbound_sasl(&mut *stream, &creds, server_name).await?;
+            crabka_client_core::outbound_sasl(&mut *stream, &to_client_creds(&creds), server_name)
+                .await
+                .map_err(|e| InterBrokerError::Sasl(e.to_string()))?;
         }
         crabka_client_core::Connection::from_stream(stream, options)
             .await
             .map_err(|e| InterBrokerError::Config(format!("Connection::from_stream: {e}")))
     }
-}
-
-// ────────────────────────────────────────────────────────────────────────
-// Outbound SASL state machine.
-// ────────────────────────────────────────────────────────────────────────
-
-async fn run_outbound_sasl<S>(
-    stream: &mut S,
-    creds: &InterBrokerCredentials,
-    server_name: &str,
-) -> Result<(), InterBrokerError>
-where
-    S: AsyncRead + AsyncWrite + Unpin + Send + ?Sized,
-{
-    // Step 1: ApiVersions. The JVM client always sends this first; we
-    //         skip it for simplicity (plan §16 step 3). The broker's
-    //         pre-auth allowlist tolerates skipping ApiVersions.
-    // Step 2: SaslHandshake with the chosen mechanism — establishes
-    //         which SASL flow the broker will run.
-    let mut corr_id: i32 = 1;
-    send_sasl_handshake(stream, creds.mechanism(), &mut corr_id).await?;
-    // Step 3: SaslAuthenticate (one round for PLAIN, two for SCRAM, three
-    //         for GSSAPI).
-    match creds {
-        InterBrokerCredentials::Plain { username, password } => {
-            send_plain_authenticate(stream, username, password, &mut corr_id).await
-        }
-        InterBrokerCredentials::Scram {
-            mechanism,
-            username,
-            password,
-        } => run_scram_client(stream, username, password, *mechanism, &mut corr_id).await,
-        InterBrokerCredentials::Gssapi {
-            keytab_path,
-            client_principal,
-            service_name,
-            kdc_url,
-        } => {
-            run_gssapi_client(
-                stream,
-                keytab_path,
-                client_principal,
-                service_name,
-                server_name,
-                kdc_url,
-                &mut corr_id,
-            )
-            .await
-        }
-    }
-}
-
-/// Send `SaslHandshakeRequest v1` with the wire name for `mechanism`,
-/// read `SaslHandshakeResponse v1`, fail if `error_code != 0`.
-///
-/// Wire framing: `SaslHandshake v1` uses the non-flexible request header
-/// (v1 — no trailing tagged-fields byte) and a non-flexible response
-/// header (v0 — bare `correlation_id`). This matches the server-side
-/// `drive_sasl_plain_session` helper in the integration test.
-async fn send_sasl_handshake<S>(
-    stream: &mut S,
-    mechanism: SaslMechanism,
-    corr_id: &mut i32,
-) -> Result<(), InterBrokerError>
-where
-    S: AsyncRead + AsyncWrite + Unpin + Send + ?Sized,
-{
-    let req = SaslHandshakeRequest {
-        mechanism: mechanism.wire_name().to_string(),
-        ..Default::default()
-    };
-    let mut body = BytesMut::new();
-    req.encode(&mut body, 1)
-        .map_err(|e| InterBrokerError::Codec(format!("SaslHandshake encode: {e}")))?;
-    let resp_bytes = round_trip(stream, API_KEY_SASL_HANDSHAKE, 1, *corr_id, false, &body).await?;
-    *corr_id += 1;
-    let mut cur: &[u8] = &resp_bytes;
-    let resp = SaslHandshakeResponse::decode(&mut cur, 1)
-        .map_err(|e| InterBrokerError::Codec(format!("SaslHandshake decode: {e}")))?;
-    if resp.error_code != 0 {
-        return Err(InterBrokerError::Sasl(format!(
-            "SaslHandshake error_code={} (mechanism={})",
-            resp.error_code,
-            mechanism.wire_name()
-        )));
-    }
-    Ok(())
-}
-
-/// Send `SaslAuthenticate v2` with PLAIN payload `\0user\0password`, read
-/// the response, fail if `error_code != 0`. PLAIN is one round-trip.
-async fn send_plain_authenticate<S>(
-    stream: &mut S,
-    user: &str,
-    pass: &str,
-    corr_id: &mut i32,
-) -> Result<(), InterBrokerError>
-where
-    S: AsyncRead + AsyncWrite + Unpin + Send + ?Sized,
-{
-    let mut payload = Vec::with_capacity(2 + user.len() + pass.len());
-    payload.push(0); // authzid (empty)
-    payload.extend_from_slice(user.as_bytes());
-    payload.push(0);
-    payload.extend_from_slice(pass.as_bytes());
-
-    let resp = send_sasl_authenticate(stream, payload, corr_id).await?;
-    if resp.error_code != 0 {
-        return Err(InterBrokerError::Sasl(format!(
-            "SaslAuthenticate(PLAIN) error_code={} error_message={:?}",
-            resp.error_code, resp.error_message
-        )));
-    }
-    Ok(())
-}
-
-/// Run the RFC 5802 SCRAM (SHA-256 or SHA-512) client state machine
-/// over two `SaslAuthenticate v2` round-trips. Verifies the
-/// server-final signature before declaring the connection
-/// authenticated.
-async fn run_scram_client<S>(
-    stream: &mut S,
-    user: &str,
-    pass: &str,
-    mechanism: SaslMechanism,
-    corr_id: &mut i32,
-) -> Result<(), InterBrokerError>
-where
-    S: AsyncRead + AsyncWrite + Unpin + Send + ?Sized,
-{
-    let mut exch = ScramClientExchange::new(user.to_string(), pass.as_bytes().to_vec(), mechanism);
-
-    // Round 1: client-first → server-first.
-    let client_first = exch
-        .client_first()
-        .map_err(|e| InterBrokerError::Sasl(format!("scram client_first: {e:?}")))?;
-    let resp1 = send_sasl_authenticate(stream, client_first, corr_id).await?;
-    if resp1.error_code != 0 {
-        return Err(InterBrokerError::Sasl(format!(
-            "SaslAuthenticate(SCRAM round 1) error_code={} error_message={:?}",
-            resp1.error_code, resp1.error_message
-        )));
-    }
-    let server_first = resp1.auth_bytes.to_vec();
-
-    // Round 2: client-final → server-final.
-    let client_final = exch
-        .step(&server_first)
-        .map_err(|e| InterBrokerError::Sasl(format!("scram client step: {e:?}")))?;
-    let resp2 = send_sasl_authenticate(stream, client_final, corr_id).await?;
-    if resp2.error_code != 0 {
-        return Err(InterBrokerError::Sasl(format!(
-            "SaslAuthenticate(SCRAM round 2) error_code={} error_message={:?}",
-            resp2.error_code, resp2.error_message
-        )));
-    }
-    // Server-final verification proves the broker holds the matching
-    // `server_key` — not just any compatible `stored_key`.
-    exch.verify_server_final(&resp2.auth_bytes)
-        .map_err(|e| InterBrokerError::Sasl(format!("server-final verify: {e:?}")))?;
-    Ok(())
-}
-
-/// Run the SASL/GSSAPI (Kerberos) client state machine over
-/// `SaslAuthenticate v2` round-trips.
-///
-/// Builds an `sspi`-backed initiator that authenticates as `client_principal`
-/// using the long-term key in `keytab_path` (no password), targeting the SPN
-/// `service_name/server_name` of the broker being dialed. `server_name` is the
-/// broker's canonical hostname (the same value used for TLS SNI), not the
-/// dialed IP — the SPN must match the service key in the broker's keytab.
-/// Drives [`GssapiClientExchange`]: the GSS context establishment
-/// (AP-REQ → AP-REP) followed by the RFC 4752 auth-only security-layer
-/// negotiation. Each non-terminal client token is sent as request
-/// `auth_bytes`; the server's reply token feeds the next step until the
-/// exchange reports `Done`.
-///
-/// The first initiator step performs the synchronous AS/TGS exchange with the
-/// KDC; subsequent steps only process tokens locally.
-async fn run_gssapi_client<S>(
-    stream: &mut S,
-    keytab_path: &Path,
-    client_principal: &str,
-    service_name: &str,
-    server_name: &str,
-    kdc_url: &str,
-    corr_id: &mut i32,
-) -> Result<(), InterBrokerError>
-where
-    S: AsyncRead + AsyncWrite + Unpin + Send + ?Sized,
-{
-    use crabka_security::gssapi::client::{ClientStep, GssapiClientExchange};
-    use crabka_security::gssapi::provider::SspiInitiator;
-
-    let target_spn = format!("{service_name}/{server_name}");
-    let keytab = keytab_path.to_string_lossy();
-    let initiator = SspiInitiator::new(&keytab, client_principal, &target_spn, kdc_url)
-        .map_err(|e| InterBrokerError::Sasl(format!("GSSAPI initiator init failed: {e}")))?;
-    let mut exchange = GssapiClientExchange::new(Box::new(initiator), GSSAPI_MAX_RECV_SIZE, None);
-
-    // Seed the exchange with no server token; this produces the AP-REQ.
-    let mut step = exchange
-        .step(None)
-        .map_err(|e| InterBrokerError::Sasl(format!("GSSAPI initiate failed: {e}")))?;
-    loop {
-        match step {
-            ClientStep::Token(token) => {
-                let resp = send_sasl_authenticate(stream, token, corr_id).await?;
-                if resp.error_code != 0 {
-                    return Err(InterBrokerError::Sasl(format!(
-                        "SaslAuthenticate(GSSAPI) error_code={} error_message={:?}",
-                        resp.error_code, resp.error_message
-                    )));
-                }
-                step = exchange
-                    .step(Some(&resp.auth_bytes))
-                    .map_err(|e| InterBrokerError::Sasl(format!("GSSAPI step failed: {e}")))?;
-            }
-            ClientStep::Done => return Ok(()),
-        }
-    }
-}
-
-/// Frame a `SaslAuthenticate v2` request carrying `auth_bytes`, send it,
-/// read the response, return the decoded `SaslAuthenticateResponse v2`.
-async fn send_sasl_authenticate<S>(
-    stream: &mut S,
-    auth_bytes: Vec<u8>,
-    corr_id: &mut i32,
-) -> Result<SaslAuthenticateResponse, InterBrokerError>
-where
-    S: AsyncRead + AsyncWrite + Unpin + Send + ?Sized,
-{
-    let req = SaslAuthenticateRequest {
-        auth_bytes: bytes::Bytes::from(auth_bytes),
-        ..Default::default()
-    };
-    let mut body = BytesMut::new();
-    req.encode(&mut body, 2)
-        .map_err(|e| InterBrokerError::Codec(format!("SaslAuthenticate encode: {e}")))?;
-    let resp_bytes =
-        round_trip(stream, API_KEY_SASL_AUTHENTICATE, 2, *corr_id, true, &body).await?;
-    *corr_id += 1;
-    let mut cur: &[u8] = &resp_bytes;
-    let resp = SaslAuthenticateResponse::decode(&mut cur, 2)
-        .map_err(|e| InterBrokerError::Codec(format!("SaslAuthenticate decode: {e}")))?;
-    Ok(resp)
-}
-
-// ────────────────────────────────────────────────────────────────────────
-// Framing helpers — Kafka request/response framing on the client side.
-// Mirrors the `round_trip` helper in `crates/broker/tests/auth_handlers.rs`.
-// ────────────────────────────────────────────────────────────────────────
-
-/// Build a `RequestHeader v1` (or v2 when `flexible`), append `body`, write
-/// the length-prefixed frame, read one response frame, strip the
-/// `ResponseHeader`. Returns the response body bytes.
-///
-/// Header rules (matching Kafka and the integration-test helper):
-/// - Request header: v1 for non-flexible, v2 for flexible (trailing 0x00
-///   tagged-fields byte).
-/// - Response header: v0 for non-flexible *and* for `ApiVersions(18)`
-///   regardless of body flexibility; v1 (`corr_id` + 0x00 tagged byte)
-///   for every other flexible response.
-async fn round_trip<S>(
-    stream: &mut S,
-    api_key: i16,
-    api_version: i16,
-    corr_id: i32,
-    flexible: bool,
-    body: &[u8],
-) -> Result<Vec<u8>, InterBrokerError>
-where
-    S: AsyncRead + AsyncWrite + Unpin + Send + ?Sized,
-{
-    let mut frame = BytesMut::with_capacity(16 + body.len());
-    // RequestHeader: api_key + version + corr_id + client_id (i16 NULLABLE_STRING).
-    frame.put_i16(api_key);
-    frame.put_i16(api_version);
-    frame.put_i32(corr_id);
-    frame.put_i16(
-        i16::try_from(OUTBOUND_CLIENT_ID.len())
-            .map_err(|_| InterBrokerError::Codec("client_id too long".into()))?,
-    );
-    frame.put_slice(OUTBOUND_CLIENT_ID.as_bytes());
-    if flexible {
-        frame.put_u8(0); // empty header tagged-fields
-    }
-    frame.put_slice(body);
-
-    stream
-        .write_u32(
-            u32::try_from(frame.len())
-                .map_err(|_| InterBrokerError::Codec("frame size exceeds u32".into()))?,
-        )
-        .await?;
-    stream.write_all(&frame).await?;
-    stream.flush().await?;
-
-    // Read length prefix then exactly that many bytes.
-    let resp_len = stream.read_u32().await?;
-    let mut resp = vec![0u8; resp_len as usize];
-    stream.read_exact(&mut resp).await?;
-
-    // Strip ResponseHeader: 4-byte corr_id, plus 1-byte tagged-fields for
-    // v1 (flexible body AND api_key != 18). ApiVersions is special-cased
-    // by the Kafka spec — its response header is always v0.
-    let mut cur = &resp[..];
-    if cur.len() < 4 {
-        return Err(InterBrokerError::Codec("response missing corr_id".into()));
-    }
-    let _resp_corr_id = cur.get_i32();
-    let uses_v1_header = flexible && api_key != 18;
-    if uses_v1_header {
-        if cur.is_empty() {
-            return Err(InterBrokerError::Codec(
-                "flexible response missing tagged-fields byte".into(),
-            ));
-        }
-        let _tagged = cur.get_u8();
-    }
-    Ok(cur.to_vec())
 }
 
 // ────────────────────────────────────────────────────────────────────────
